LPS-53386 Simplify arquillian-bridge to be just arquillian-junit-bridge
