--- modules/test/arquillian-bridge/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Arquillian Bridge
-Bundle-SymbolicName: com.liferay.arquillian.bridge
-Bundle-Version: 1.0.0
-Include-Resource: classes
