--- modules/test/arquillian-bridge/src/config.properties
@@ -1,45 +1,0 @@
-##
-## DO NOT EDIT THIS FILE.
-##
-## To update the properties of this file, create a separate properties file
-## named "config-ext.properties" with the properties to overwrite.
-##
-
-bundle.classpath.files=\
-    org.jboss.arquillian.container,arquillian-container-impl-base,1.1.6.Final;\
-    org.jboss.arquillian.container,arquillian-container-spi,1.1.6.Final;\
-    org.jboss.arquillian.container,arquillian-container-spi,1.1.6.Final;\
-    org.jboss.arquillian.container,arquillian-container-test-api,1.1.6.Final;\
-    org.jboss.arquillian.container,arquillian-container-test-impl-base,1.1.6.Final;\
-    org.jboss.arquillian.container,arquillian-container-test-spi,1.1.6.Final;\
-    org.jboss.arquillian.core,arquillian-core-api,1.1.6.Final;\
-    org.jboss.arquillian.core,arquillian-core-impl-base,1.1.6.Final;\
-    org.jboss.arquillian.core,arquillian-core-spi,1.1.6.Final;\
-    org.jboss.arquillian.junit,arquillian-junit-core,1.1.6.Final;\
-    org.jboss.arquillian.junit,arquillian-junit-container,1.1.6.Final;\
-    org.jboss.arquillian.protocol,arquillian-protocol-jmx,1.1.6.Final;\
-    org.jboss.arquillian.test,arquillian-test-api,1.1.6.Final;\
-    org.jboss.arquillian.test,arquillian-test-impl-base,1.1.6.Final;\
-    org.jboss.arquillian.test,arquillian-test-spi,1.1.6.Final;\
-    org.jboss.shrinkwrap,shrinkwrap-api,1.1.2;\
-    org.jboss.shrinkwrap,shrinkwrap-impl-base,1.1.2;\
-    org.jboss.shrinkwrap,shrinkwrap-spi,1.1.2
-
-bundle.import.exclude.packages=\
-    com.liferay.arquillian.bridge.junit;\
-    org.jboss.arquillian
-
-bundle.import.files=\
-    junit,junit,4.12;\
-    org.hamcrest,hamcrest-core,1.3
-
-bundle.import.packages=\
-    org.osgi.framework;\
-    javax.management
-
-dump.generated.archives=false
-
-jmx.url=service:jmx:rmi:///jndi/rmi://localhost:8099/jmxrmi
-jmx.user=
-jmx.password=
-jmx.timout=30
