+++ modules/test/arquillian-extension-junit-bridge/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Arquillian Extension JUnit Bridge
+Bundle-SymbolicName: com.liferay.arquillian.extension.junit.bridge
+Bundle-Version: 1.0.0
+Include-Resource: classes
