+++ modules/test/arquillian-extension-junit-bridge/src/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,0 +1,1 @@
+com.liferay.arquillian.extension.junit.bridge.LiferayArquillianJUnitBridgeExtension
