+++ modules/test/arquillian-extension-junit-bridge/src/com/liferay/arquillian/extension/junit/bridge/observer/JUnitBridgeObserver.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.arquillian.bridge.junit.observer;
+package com.liferay.arquillian.extension.junit.bridge.observer;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
